html/interaction/focus/sequential-focus-navigation-and-the-tabindex-attribute/focus-tabindex-order-shadow-negative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focus - the sequential focus navigation order with shadow dom and negative tabindex in shadow scope</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#sequential-focus-navigation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/shadow-utils.js"></script>
<body>
<script>
// Structure:
// <div #aboveHost tabindex=0>
// <div #host tabindex=0>
//    #shadowRoot
//      <div #aboveSlot tabindex=-1>
//      <slot #slotAbove tabindex=-1>
//        (slotted) <div #slottedAbove tabindex=-1>
//      <slot #slotBelow tabindex=-1>
//        (slotted) <div #slottedBelow tabindex=-1>
//      <div #belowSlot tabindex=-1>
// <div #belowHost tabindex=0>

async_test(async (t) => {
  let elementsInFlatTreeOrder;
  let [aboveHost, host, aboveSlot, slotAbove, slottedAbove, slotBelow, slottedBelow, belowSlot, belowHost] =
    elementsInFlatTreeOrder = prepareDOM(document.body, false);
  setTabIndex(elementsInFlatTreeOrder, -1);
  setTabIndex([aboveHost, host, belowHost], 0);
  resetFocus();
  const expectedOrder = [aboveHost, host, belowHost]; 
  for (let el of expectedOrder) {
    await navigateFocusForward();
    let focused = host.shadowRoot.activeElement ? host.shadowRoot.activeElement : document.activeElement;
    t.step(() => { assert_equals(focused, el) });
  }
  t.done();
}, "Order when all tabindex=0 is and delegatesFocus = false");
</script>
</body>
